Rename study modules to flow/run with dedicated bin
