+++ CMakeLists.txt
@@ -36,4 +36,4 @@ add_subdirectory(src/syst)
 add_subdirectory(src/core)
 add_subdirectory(src/plug)
 add_subdirectory(src/tests)
-add_subdirectory(src/study)
+add_subdirectory(src/run)
