+++ src/run/CMakeLists.txt
@@ -33,3 +33,6 @@ target_link_libraries(study_all PRIVATE
   TBB::tbb
   TMVA
   dl)
+
+set_target_properties(study_topo_score study_all PROPERTIES
+  RUNTIME_OUTPUT_DIRECTORY "${CMAKE_SOURCE_DIR}/bin")
